requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
google-generativeai==0.8.6
pinecone==8.1.0
pdfplumber==0.11.9
langchain-text-splitters==1.1.1
tiktoken==0.12.0
python-dotenv==1.2.2
requests==2.33.0
tenacity==9.1.4
tqdm==4.67.3
rich==14.3.3
python-docx==1.1.2
python-pptx==1.0.2
